# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Aida Ramezani Art Portfolio & Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type ArtworkStatus = typeof ArtworkStatus[keyof typeof ArtworkStatus];


export const ArtworkStatus = {
  AVAILABLE: 'AVAILABLE',
  SOLD: 'SOLD',
  RESERVED: 'RESERVED',
} as const;

export interface Artwork {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  medium: string;
  /** @nullable */
  sizeInches?: string | null;
  year: number;
  imageUrl: string;
  /** @nullable */
  priceCents?: number | null;
  currency?: string;
  status: ArtworkStatus;
  category: string;
  availableAsPrint?: boolean;
  /** @nullable */
  printfulProductId?: string | null;
  createdAt: string;
}

export type ArtworkUpdateStatus = typeof ArtworkUpdateStatus[keyof typeof ArtworkUpdateStatus];


export const ArtworkUpdateStatus = {
  AVAILABLE: 'AVAILABLE',
  SOLD: 'SOLD',
  RESERVED: 'RESERVED',
} as const;

export interface ArtworkUpdate {
  title?: string;
  description?: string;
  status?: ArtworkUpdateStatus;
  priceCents?: number;
  category?: string;
  availableAsPrint?: boolean;
  printfulProductId?: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface ArtworkStats {
  total: number;
  available: number;
  sold: number;
  reserved: number;
  byCategory: CategoryCount[];
}

export interface Event {
  id: number;
  title: string;
  description: string;
  dateTime: string;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  signupCount?: number;
  createdAt: string;
}

export interface EventInput {
  title: string;
  description: string;
  dateTime: string;
  location: string;
  imageUrl?: string;
}

export interface EventSignup {
  id: number;
  eventId: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface EventSignupInput {
  name: string;
  email: string;
}

export interface NewsletterSubscriber {
  id: number;
  email: string;
  createdAt: string;
}

export interface NewsletterInput {
  email: string;
  name?: string;
}

export type CheckoutRequestType = typeof CheckoutRequestType[keyof typeof CheckoutRequestType];


export const CheckoutRequestType = {
  original: 'original',
  print: 'print',
} as const;

export interface CheckoutRequest {
  type: CheckoutRequestType;
  artworkId: number;
  printSize?: string;
  customerEmail?: string;
}

export interface CheckoutResponse {
  checkoutUrl: string;
  /** @nullable */
  transactionId?: string | null;
}

export interface WebhookAck {
  received: boolean;
}

export type ListArtworksParams = {
status?: ListArtworksStatus;
category?: string;
limit?: number;
};

export type ListArtworksStatus = typeof ListArtworksStatus[keyof typeof ListArtworksStatus];


export const ListArtworksStatus = {
  AVAILABLE: 'AVAILABLE',
  SOLD: 'SOLD',
  RESERVED: 'RESERVED',
} as const;

export type ListEventsParams = {
upcoming?: boolean;
};