# lib/api-zod/src/generated/types/artworkStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Aida Ramezani Art Portfolio & Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';

export interface ArtworkStats {
  total: number;
  available: number;
  sold: number;
  reserved: number;
  byCategory: CategoryCount[];
}