# lib/api-zod/src/generated/types/artworkUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Aida Ramezani Art Portfolio & Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { ArtworkUpdateStatus } from './artworkUpdateStatus';

export interface ArtworkUpdate {
  title?: string;
  description?: string;
  status?: ArtworkUpdateStatus;
  priceCents?: number;
  category?: string;
  availableAsPrint?: boolean;
  printfulProductId?: string;
}