# lib/api-zod/src/generated/types/newsletterSubscriber.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Aida Ramezani Art Portfolio & Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface NewsletterSubscriber {
  id: number;
  email: string;
  createdAt: Date;
}